Break down profile into post and photo tab cmpnts
