+++ src/routes/Profile.tsx
@@ -3,30 +3,23 @@ import { useParams } from 'react-router-dom'
 
 import { getUserDocFromUsername } from '../utils/firebase/firebase-config'
 
-import { MdRepeat } from 'react-icons/md'
-
 import { MenuContext } from '../contexts/MenuContext'
 import { UserContext } from '../contexts/User'
-import { UserPostsContext } from '../contexts/UserPosts'
 import { ModalContext } from '../contexts/ModalContext'
 
-import { IUserRepost } from '../contexts/UserPosts'
-
 import FeedContainer from '../components/Feed/FeedContainer'
-import Post from '../components/Posts/Post'
-import LoadingSpinner from '../components/LoadingSpinner'
 import ProfileBanner from '../components/Profile/ProfileBanner'
-import CreatePost from '../components/CreatePost/CreatePost'
+import ProfilePostsTab from '../components/Profile/ProfilePostsTab'
+import LoadingSpinner from '../components/LoadingSpinner'
+import ProfilePhotosTab from '../components/Profile/ProfilePhotosTab'
 
 function Profile() {
   const { setIsOpen } = useContext(MenuContext)
   const { currentAuthUser, currentUserDoc } = useContext(UserContext)
-  const { userPosts } = useContext(UserPostsContext)
   const { setModalIsOpen } = useContext(ModalContext)
 
   const [profileUserDoc, setProfileUserDoc] = useState(null)
-  const [profilePosts, setProfilePosts] = useState(null)
-  const [profileReposts, setProfileReposts] = useState(null)
+  const [tab, setTab] = useState('posts')
 
   const username = useParams().username.toLowerCase()
 
@@ -41,67 +34,46 @@ function Profile() {
     })
   }, [username, currentUserDoc])
 
-  useEffect(() => {
-    setProfileReposts(null)
-    if (userPosts && profileUserDoc) {
-      if (profileUserDoc.reposts) {
-        let repostOriginals = []
-        profileUserDoc.reposts.map((repost: IUserRepost) => {
-          const original = userPosts.find((post) => post.postId === repost.postId)
-          repostOriginals.push({ ...original, isRepost: true, repostTimestamp: repost.timestamp })
-        })
-        setProfileReposts(repostOriginals)
-      }
-      setProfilePosts(
-        userPosts.filter((post) => {
-          return post.username === profileUserDoc.username
-        })
-      )
-    } else return
-  }, [userPosts, profileUserDoc])
+  const postsTabClickHandler = () => {
+    setTab('posts')
+  }
+
+  const photosTabClickHandler = () => {
+    setTab('photos')
+  }
 
   if (profileUserDoc) {
-    let allProfilePosts = profileReposts ? profilePosts.concat(profileReposts) : profilePosts
-    if (allProfilePosts) {
-      const allProfilePostsSorted = allProfilePosts.sort((a: any, b: any) => {
-        if (a.isRepost && b.isRepost) {
-          return new Date(a.repostTimestamp.valueOf() - new Date(b.repostTimestamp.valueOf()))
-        } else if (a.isRepost) {
-          return new Date(a.repostTimestamp.valueOf() - new Date(b.timestamp.valueOf()))
-        } else if (b.isRepost) {
-          return new Date(a.repostTimestamp.valueOf() - new Date(b.timestamp.valueOf()))
-        } else {
-          return new Date(a.timestamp).valueOf() - new Date(b.timestamp).valueOf()
-        }
-      })
-      allProfilePosts = allProfilePostsSorted
-    }
+    console.log(profileUserDoc)
     return (
       <FeedContainer>
         <ProfileBanner currentAuthUser={currentAuthUser} profileUserDoc={profileUserDoc} />
-        <div className='text-center text-lg pt-5 pb-3 border-b border-l border-r border-slate-700'>
-          <span>
-            <b>Posts</b>
-          </span>
+        <div className='flex text-lg pt-5 pb-3 border-b border-l border-r border-slate-700'>
+          <div className='w-1/2 text-center'>
+            <span
+              className={`${
+                tab === 'posts' && 'underline'
+              } underline-offset-8 decoration-cyan-500 hover:cursor-pointer`}
+              onClick={postsTabClickHandler}
+            >
+              <b>Posts</b>
+            </span>
+          </div>
+          <div className='w-1/2 text-center'>
+            <span
+              className={`${
+                tab === 'photos' && 'underline'
+              } underline-offset-8 decoration-cyan-500 hover:cursor-pointer`}
+              onClick={photosTabClickHandler}
+            >
+              <b>Photos</b>
+            </span>
+          </div>
         </div>
-        {currentAuthUser.uid === profileUserDoc.uid && <CreatePost />}
-        {allProfilePosts ? (
-          allProfilePosts.reverse().map((post: any) => {
-            if (post.isRepost) {
-              return (
-                <div>
-                  <div className='pl-5 pt-2 border-l border-r border-slate-700 text-gray-500'>
-                    <MdRepeat className='inline mr-5 text-lg mt-0' />
-                    <span className='text-sm'>{currentUserDoc.displayName} Reposted</span>
-                  </div>
-                  <Post key={post.postId} post={post} postPage={false} />
-                </div>
-              )
-            }
-            return <Post key={post.postId} post={post} postPage={false} />
-          })
-        ) : (
-          <LoadingSpinner />
+        {tab === 'posts' && (
+          <ProfilePostsTab profileUserDoc={profileUserDoc} currentAuthUser={currentAuthUser} />
+        )}
+        {tab === 'photos' && (
+          <ProfilePhotosTab profileUserDoc={profileUserDoc} currentAuthUser={currentAuthUser} />
         )}
       </FeedContainer>
     )
